requirements.txt
fastapi==0.38.1
newsapi-python==0.2.5
praw==6.3.1
pytest==5.1.3
uvicorn==0.9.0
